pyproject.toml
[project]
name = "ankibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.2",
    "boto3>=1.36.16",
    "genanki>=0.13.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==20.8",
]
